Add boxShadowSize prop to Box
