+++ packages/core/src/utils/index.js
@@ -2,3 +2,4 @@ export * from './utils'
 export { default as createColorStyles } from './createColorStyles'
 export { default as createTextStyles } from './createTextStyles'
 export { default as createTheme } from './createTheme'
+export { default as boxShadow } from './boxShadow'
